requirements.txt
# docker/requirements.txt
requests
kafka-python==2.0.2
six==1.16.0
minio==7.2.9
psycopg2-binary==2.9.10
pandas==2.2.3
pyarrow==15.0.2

pyspark==3.5.3

torch==2.5.1
mlflow==2.14.1

boto3==1.34.162
